Moved common migration logic to the migration module
