+++ components/merkledb/examples/manual_migration.rs
@@ -25,44 +25,14 @@
 //!   to its intended place and removes the old data marked for removal.
 
 use exonum_merkledb::{
-    access::Prefixed,
-    migration::{flush_migration, Migration},
-    Database, Fork, ObjectHash, ReadonlyFork,
+    access::Prefixed, migration::Migration, Database, Fork, ObjectHash, ReadonlyFork,
 };
 
-use migration::{check_data_after_flush, check_data_before_flush, create_initial_data, v1, v2};
+use std::sync::Arc;
+
+use migration::{perform_migration, v1, v2};
 
 mod migration;
-
-fn create_migration(fork: &Fork) {
-    println!("\nStarted migration");
-
-    {
-        let new_data = Migration::new("test", fork.clone());
-        let old_data = Prefixed::new("test", fork.readonly());
-
-        let old_schema = v1::Schema::new(old_data);
-        let mut new_schema = v2::Schema::new(new_data);
-
-        // Move `ticker` and `divisibility` to `config`.
-        let config = v2::Config {
-            ticker: old_schema.ticker.get().unwrap(),
-            divisibility: old_schema.divisibility.get().unwrap_or(0),
-        };
-        new_schema.config.set(config);
-        // Mark these two indexes for removal.
-        new_data.create_tombstone("ticker");
-        new_data.create_tombstone("divisibility");
-    }
-
-    // Migrate wallets using schema:
-    // `Wallet::public_key` field will be removed.
-    // `Wallet::history_hash` field will be added.
-    // Wallets and history from username Eve will be removed.
-    let new_data = Migration::new("test", fork);
-    let old_data = Prefixed::new("test", fork.readonly());
-    migrate_wallets(new_data, old_data);
-}
 
 /// Provides migration of wallets with schema.
 fn migrate_wallets(new_data: Migration<&Fork>, old_data: Prefixed<ReadonlyFork>) {
@@ -94,51 +64,40 @@ fn migrate_wallets(new_data: Migration<&Fork>, old_data: Prefixed<ReadonlyFork>)
     }
 }
 
-fn main() {
-    // Creating a temporary DB and filling it with some data.
-    let db = create_initial_data();
-
+fn manual_migration(db: Arc<dyn Database>) {
+    // Create fork to apply changes to it.
     let fork = db.fork();
+
     {
-        // State before migration.
+        let new_data = Migration::new("test", &fork);
         let old_data = Prefixed::new("test", fork.readonly());
-        let old_schema = v1::Schema::new(old_data.clone());
-        println!("Before migration:");
-        old_schema.print_wallets();
+
+        let old_schema = v1::Schema::new(old_data);
+        let mut new_schema = v2::Schema::new(new_data);
+
+        // Move `ticker` and `divisibility` to `config`.
+        let config = v2::Config {
+            ticker: old_schema.ticker.get().unwrap(),
+            divisibility: old_schema.divisibility.get().unwrap_or(0),
+        };
+        new_schema.config.set(config);
+        // Mark these two indexes for removal.
+        new_data.create_tombstone("ticker");
+        new_data.create_tombstone("divisibility");
     }
 
-    // Migrate the data.
-    create_migration(&fork);
+    // Migrate wallets using schema:
+    // `Wallet::public_key` field will be removed.
+    // `Wallet::history_hash` field will be added.
+    // Wallets and history from username Eve will be removed.
+    let new_data = Migration::new("test", &fork);
+    let old_data = Prefixed::new("test", fork.readonly());
+    migrate_wallets(new_data, old_data);
+
     // Merge patch with migrated data.
     db.merge(fork.into_patch()).unwrap();
+}
 
-    // At this point the old data and new data are still present in the storage,
-    // but new data is in the unmerged form.
-
-    // Check that DB contains old and new data.
-    let snapshot = db.snapshot();
-    check_data_before_flush(&snapshot);
-
-    // Finalize the migration by calling `flush_migration`.
-    let mut fork = db.fork();
-    flush_migration(&mut fork, "test");
-
-    // At this point the new indexes have replaced the old ones in the fork.
-    // And indexes are aggregated in the default namespace.
-
-    // Check that indexes are updated.
-    let patch = fork.into_patch();
-    check_data_after_flush(&patch);
-
-    // When the patch is merged, the situation remains the same.
-    db.merge(patch).unwrap();
-
-    // Check that data was updated after merge.
-    let snapshot = db.snapshot();
-    check_data_after_flush(&snapshot);
-
-    // State after migration.
-    let schema = v2::Schema::new(Prefixed::new("test", &snapshot));
-    println!("After migration:");
-    schema.print_wallets();
+fn main() {
+    perform_migration(manual_migration);
 }
