+++ components/merkledb/examples/migration_with_iter_loop.rs
@@ -26,45 +26,13 @@
 //! 4. We complete the migration by calling `flush_migration`. This moves the migrated data
 //!   to its intended place and removes the old data marked for removal.
 
-use exonum_merkledb::{
-    access::Prefixed,
-    migration::{flush_migration, MigrationHelper},
-    Database, ObjectHash, Result as DbResult,
-};
+use exonum_merkledb::{migration::MigrationHelper, Database, ObjectHash, Result as DbResult};
 
 use std::sync::Arc;
 
-use migration::{check_data_after_flush, check_data_before_flush, create_initial_data, v1, v2};
+use migration::{perform_migration, v1, v2};
 
 mod migration;
-
-fn create_migration(helper: &mut MigrationHelper) -> DbResult<()> {
-    println!("\nStarted migration");
-
-    {
-        let new_data = helper.new_data();
-        let old_data = helper.old_data();
-
-        let old_schema = v1::Schema::new(old_data);
-        let mut new_schema = v2::Schema::new(new_data);
-
-        // Move `ticker` and `divisibility` to `config`.
-        let config = v2::Config {
-            ticker: old_schema.ticker.get().unwrap(),
-            divisibility: old_schema.divisibility.get().unwrap_or(0),
-        };
-        new_schema.config.set(config);
-        // Mark these two indexes for removal.
-        new_data.create_tombstone("ticker");
-        new_data.create_tombstone("divisibility");
-    }
-
-    // Migrate wallets using `MigrationHelper::iter_loop`:
-    // `Wallet::public_key` field will be removed.
-    // `Wallet::history_hash` field will be added.
-    // Wallets and history from username Eve will be removed.
-    migrate_wallets(helper)
-}
 
 /// Provides migration of wallets with `MigrationHelper::iter_loop`.
 /// `iter_loop` is designed to allow to merge changes to the database from time to time,
@@ -103,59 +71,38 @@ fn migrate_wallets(helper: &mut MigrationHelper) -> DbResult<()> {
     })
 }
 
-fn main() {
-    // Creating a temporary DB and filling it with some data.
-    let db: Arc<dyn Database> = Arc::new(create_initial_data());
-
+fn migration_with_iter_loop(db: Arc<dyn Database>) {
     // Creating helper to perform migration.
     let mut helper = MigrationHelper::new(db.clone(), "test");
 
     {
-        // State before migration.
-        let old_schema = v1::Schema::new(helper.old_data());
-        println!("Before migration:");
-        old_schema.print_wallets();
+        let new_data = helper.new_data();
+        let old_data = helper.old_data();
+
+        let old_schema = v1::Schema::new(old_data);
+        let mut new_schema = v2::Schema::new(new_data);
+
+        // Move `ticker` and `divisibility` to `config`.
+        let config = v2::Config {
+            ticker: old_schema.ticker.get().unwrap(),
+            divisibility: old_schema.divisibility.get().unwrap_or(0),
+        };
+        new_schema.config.set(config);
+        // Mark these two indexes for removal.
+        new_data.create_tombstone("ticker");
+        new_data.create_tombstone("divisibility");
     }
 
-    // Migrate the data.
-    if let Err(err) = create_migration(&mut helper) {
-        println!("Migration failed: {}", err);
-        return;
-    }
+    // Migrate wallets using schema:
+    // `Wallet::public_key` field will be removed.
+    // `Wallet::history_hash` field will be added.
+    // Wallets and history from username Eve will be removed.
+    migrate_wallets(&mut helper).expect("Wallet migration failed.");
 
     // Call `MigrationHelper::finish` to merge changes to the database.
-    if let Err(err) = helper.finish() {
-        println!("Migration finish failed: {}", err);
-        return;
-    }
+    helper.finish().expect("Migration finish failed.");
+}
 
-    // At this point the old data and new data are still present in the storage,
-    // but new data is in the unmerged form.
-
-    // Check that DB contains old and new data.
-    let snapshot = db.snapshot();
-    check_data_before_flush(&snapshot);
-
-    // Finalize the migration by calling `flush_migration`.
-    let mut fork = db.fork();
-    flush_migration(&mut fork, "test");
-
-    // At this point the new indexes have replaced the old ones in the fork.
-    // And indexes are aggregated in the default namespace.
-
-    // Check that indexes are updated.
-    let patch = fork.into_patch();
-    check_data_after_flush(&patch);
-
-    // When the patch is merged, the situation remains the same.
-    db.merge(patch).unwrap();
-
-    // Check that data was updated after merge.
-    let snapshot = db.snapshot();
-    check_data_after_flush(&snapshot);
-
-    // State after migration.
-    let schema = v2::Schema::new(Prefixed::new("test", &snapshot));
-    println!("After migration:");
-    schema.print_wallets();
+fn main() {
+    perform_migration(migration_with_iter_loop);
 }
